Refactoring unit tests
