+++ tests/core/test_schema.py
@@ -1,0 +1,15 @@
+# pyright: reportPrivateImportUsage=false, reportPrivateUsage=false, reportUnknownParameterType=false, reportMissingParameterType=false, reportAny=false, reportUnknownMemberType=false, reportUnknownArgumentType=false, reportArgumentType=false, reportFunctionMemberAccess=false, reportUnknownVariableType=false, reportUnusedParameter=false
+
+from dbt_osmosis.core.schema.parser import create_yaml_instance
+
+
+def test_create_yaml_instance_settings():
+    """
+    Quick check that create_yaml_instance returns a configured YAML object with custom indenting.
+    """
+    y = create_yaml_instance(indent_mapping=4, indent_sequence=2, indent_offset=0)
+    assert y.map_indent == 4
+    assert y.sequence_indent == 2
+    assert y.sequence_dash_offset == 0
+    assert y.width == 100  # default
+    assert y.preserve_quotes is False
